:recycle: refactory code, to suit to pull request
