+++ src/Recurrence/Aggregates/Invoice.php
@@ -36,5 +36,4 @@ class Invoice extends AbstractEntity
             'subscriptionId' => $this->getSubscriptionId()
         ];
     }
-
 }
