+++ src/Recurrence/Factories/SubscriptionFactory.php
@@ -30,7 +30,9 @@ class SubscriptionFactory implements FactoryInterface
         $subscription->setStatus(SubscriptionStatus::{$postData['status']}());
         $subscription->setInstallments($postData['installments']);
         $subscription->setPaymentMethod(PaymentMethod::{$postData['payment_method']}());
-        $subscription->setIntervalType(IntervalValueObject::{$postData['interval']}($postData['interval_count']));
+        $subscription->setIntervalType(
+            IntervalValueObject::{$postData['interval']}($postData['interval_count'])
+        );
 
         $subscription->setMundipaggId(new SubscriptionId($postData['id']));
 
@@ -78,7 +80,9 @@ class SubscriptionFactory implements FactoryInterface
         $subscription->setStatus(SubscriptionStatus::{$dbData['status']}());
         $subscription->setInstallments($dbData['installments']);
         $subscription->setPaymentMethod(PaymentMethod::{$dbData['payment_method']}());
-        $subscription->setIntervalType(IntervalValueObject::{$dbData['interval_type']}($dbData['interval_count']));
+        $subscription->setIntervalType(
+            IntervalValueObject::{$dbData['interval_type']}($dbData['interval_count'])
+        );
 
         $subscription->setPlatformOrder($this->getPlatformOrder($dbData['code']));
